Fix Qulice vulnerability

+++ src/main/java/org/polystat/Program.java
@@ -70,10 +70,8 @@ public final class Program implements Func<String, XML> {
         final String name = parts[1];
         final Path xml = this.temp.resolve(String.format("%s.xml", name)).toAbsolutePath();
         final Path src = this.sources.resolve(String.format("%s.eo", name)).toAbsolutePath();
-        if (!src.toFile().exists()) {
-            if (xml.toFile().exists()) {
-                xml.toFile().delete();
-            }
+        if (!src.toFile().exists() && xml.toFile().exists()) {
+            xml.toFile().delete();
         }
         if (
             !xml.toFile().exists()
